+++ src/main/java/skizlit/jcmods/common/blocks/BlockBasic.java
@@ -1,4 +1,4 @@
-package skizlit.jcmods.blocks;
+package skizlit.jcmods.common.blocks;
 
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
