+++ src/main/java/skizlit/jcmods/common/util/IModelled.java
@@ -1,4 +1,4 @@
-package skizlit.jcmods.util;
+package skizlit.jcmods.common.util;
 
 public interface IModelled {
 	public void registerModels();
